+++ packages/payment-proxy-client/src/App.tsx
@@ -18,7 +18,7 @@ import {
 const QUERY_KEY = "rpcUrl";
 
 import "./App.css";
-import { fetchFile } from "./file-helpers";
+import { fetchFile } from "./file";
 
 const provider = "0xbbb676f9cff8d242e9eac39d063848807d3d1d94";
 const hub = "0x111a00868581f73ab42feef67d235ca09ca1e8db";
